requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scipy>=1.11.0
scikit-learn>=1.3.0
Pillow>=10.0.0
streamlit>=1.28.0
imutils>=0.5.4
